+++ Core.lua
@@ -1,7 +1,7 @@
-local ADDON_NAME, CombatState = ...
+local ADDON_NAME, CombatCue = ...
 
 local eventFrame = CreateFrame("Frame")
-local prefix = "|cff33ff99CombatState|r:"
+local prefix = "|cff33ff99CombatCue|r:"
 
 local function Print(message)
     DEFAULT_CHAT_FRAME:AddMessage(prefix .. " " .. message)
@@ -10,18 +10,18 @@ end
 local function OnEvent(_, event, addonName)
     if event == "ADDON_LOADED" then
         if addonName == ADDON_NAME then
-            CombatState:EnsureDB()
-            CombatState:ApplySettings()
-            Print(CombatState.L.loaded)
+            CombatCue:EnsureDB()
+            CombatCue:ApplySettings()
+            Print(CombatCue.L.loaded)
         end
     elseif event == "PLAYER_REGEN_DISABLED" then
-        local color = CombatState:GetCombatColor("enter")
+        local color = CombatCue:GetCombatColor("enter")
 
-        CombatState:ShowAlert(CombatStateDB.enterCombatMessage, color.r, color.g, color.b, color.a)
+        CombatCue:ShowAlert(CombatCueDB.enterCombatMessage, color.r, color.g, color.b, color.a)
     elseif event == "PLAYER_REGEN_ENABLED" then
-        local color = CombatState:GetCombatColor("leave")
+        local color = CombatCue:GetCombatColor("leave")
 
-        CombatState:ShowAlert(CombatStateDB.leaveCombatMessage, color.r, color.g, color.b, color.a)
+        CombatCue:ShowAlert(CombatCueDB.leaveCombatMessage, color.r, color.g, color.b, color.a)
     end
 end
 
@@ -30,9 +30,10 @@ eventFrame:RegisterEvent("PLAYER_REGEN_DISABLED")
 eventFrame:RegisterEvent("PLAYER_REGEN_ENABLED")
 eventFrame:SetScript("OnEvent", OnEvent)
 
-SLASH_COMBATSTATE1 = "/combatstate"
-SLASH_COMBATSTATE2 = "/cs"
+SLASH_COMBATCUE1 = "/combatcue"
+SLASH_COMBATCUE2 = "/cc"
+SLASH_COMBATCUE3 = "/cs"
 
-SlashCmdList.COMBATSTATE = function()
-    CombatState:ToggleConfig()
+SlashCmdList.COMBATCUE = function()
+    CombatCue:ToggleConfig()
 end